+++ src/roswire/proxy/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from .bag import BagRecorder, BagPlayer
 from .catkin import CatkinInterface, CatkinTools, CatkinMake
-from .container import Container, ContainerManager
 from .roscore import ROSCore
 from .roslaunch import ROSLaunchManager
 from .state import SystemState